Add copy filter button to UI. Change Component section to Search Field. Bug fixes.
Add a new button to duplicate the current filter directly after itself. Copies all the values from the current filter into the new filter.
The Menu used for components was running out of space in large projects (only 1000 items are allowed). This switched that to a Popup Search Field that makes suggestions based on what the user is typing.
Miscellaneous bug fixes and clean up relating to the UI changes.